Units of measure fix launch_pre_exp.py
The code to calculate the pre exponential factor uses the Boltzmann and Planck's constants in Joules, and therefore the mass needs to be in kg in order for the units of measure to line up. 
On the previous version the mass was on atomic units, this commit fixes this issue, as well as an error in the formulas. 
The new formulas written where taken from Ligterink and Minissale paper titled "Overview of desorption parameters of volatile and complex organic molecules" from 2023